style: стилизация в соответствии с предупреждениями линтеров [задачи 9, 10]
