--- app/src/main/java/ru/practicum/android/diploma/presentation/favorites/view_model/FavoritesViewModel.kt
@@ -1,5 +1,0 @@
-package ru.practicum.android.diploma.presentation.favorites.view_model
-
-import androidx.lifecycle.ViewModel
-
-class FavoritesViewModel : ViewModel()
