+++ app/src/main/java/ru/practicum/android/diploma/presentation/favorites/viewmodel/FavoritesViewModel.kt
@@ -1,0 +1,5 @@
+package ru.practicum.android.diploma.presentation.favorites.viewmodel
+
+import androidx.lifecycle.ViewModel
+
+class FavoritesViewModel : ViewModel()
