--- app/src/main/java/ru/practicum/android/diploma/presentation/filtration/choose_country/view_model/ChooseCountryViewModel.kt
@@ -1,5 +1,0 @@
-package ru.practicum.android.diploma.presentation.filtration.choose_country.view_model
-
-import androidx.lifecycle.ViewModel
-
-class ChooseCountryViewModel : ViewModel()
