--- app/src/main/java/ru/practicum/android/diploma/presentation/filtration/choose_region/view_model/ChooseRegionViewModel.kt
@@ -1,5 +1,0 @@
-package ru.practicum.android.diploma.presentation.filtration.choose_region.view_model
-
-import androidx.lifecycle.ViewModel
-
-class ChooseRegionViewModel : ViewModel()
