+++ app/src/main/java/ru/practicum/android/diploma/presentation/filtration/country/viewmodel/ChooseCountryViewModel.kt
@@ -1,0 +1,5 @@
+package ru.practicum.android.diploma.presentation.filtration.country.viewmodel
+
+import androidx.lifecycle.ViewModel
+
+class ChooseCountryViewModel : ViewModel()
