+++ app/src/main/java/ru/practicum/android/diploma/presentation/filtration/industry/viewmodel/IndustrySelectionViewModel.kt
@@ -1,0 +1,5 @@
+package ru.practicum.android.diploma.presentation.filtration.industry.viewmodel
+
+import androidx.lifecycle.ViewModel
+
+class IndustrySelectionViewModel : ViewModel()
