--- app/src/main/java/ru/practicum/android/diploma/presentation/filtration/industry_selection/view_model/IndustrySelectionViewModel.kt
@@ -1,5 +1,0 @@
-package ru.practicum.android.diploma.presentation.filtration.industry_selection.view_model
-
-import androidx.lifecycle.ViewModel
-
-class IndustrySelectionViewModel : ViewModel()
