--- app/src/main/java/ru/practicum/android/diploma/presentation/filtration/place_of_work/view_model/PlaceOfWorkViewModel.kt
@@ -1,5 +1,0 @@
-package ru.practicum.android.diploma.presentation.filtration.place_of_work.view_model
-
-import androidx.lifecycle.ViewModel
-
-class PlaceOfWorkViewModel : ViewModel()
