+++ app/src/main/java/ru/practicum/android/diploma/presentation/filtration/region/viewmodel/ChooseRegionViewModel.kt
@@ -1,0 +1,5 @@
+package ru.practicum.android.diploma.presentation.filtration.region.viewmodel
+
+import androidx.lifecycle.ViewModel
+
+class ChooseRegionViewModel : ViewModel()
