--- app/src/main/java/ru/practicum/android/diploma/presentation/filtration/view_model/FiltrationViewModel.kt
@@ -1,5 +1,0 @@
-package ru.practicum.android.diploma.presentation.filtration.view_model
-
-import androidx.lifecycle.ViewModel
-
-class FiltrationViewModel : ViewModel()
