+++ app/src/main/java/ru/practicum/android/diploma/presentation/filtration/viewmodel/FiltrationViewModel.kt
@@ -1,0 +1,5 @@
+package ru.practicum.android.diploma.presentation.filtration.viewmodel
+
+import androidx.lifecycle.ViewModel
+
+class FiltrationViewModel : ViewModel()
