+++ app/src/main/java/ru/practicum/android/diploma/presentation/filtration/workplace/viewmodel/PlaceOfWorkViewModel.kt
@@ -1,0 +1,5 @@
+package ru.practicum.android.diploma.presentation.filtration.workplace.viewmodel
+
+import androidx.lifecycle.ViewModel
+
+class PlaceOfWorkViewModel : ViewModel()
