--- app/src/main/java/ru/practicum/android/diploma/presentation/job_search/view_model/JobSearchViewModel.kt
@@ -1,5 +1,0 @@
-package ru.practicum.android.diploma.presentation.job_search.view_model
-
-import androidx.lifecycle.ViewModel
-
-class JobSearchViewModel : ViewModel()
