+++ app/src/main/java/ru/practicum/android/diploma/presentation/search/viewmodel/JobSearchViewModel.kt
@@ -1,0 +1,5 @@
+package ru.practicum.android.diploma.presentation.search.viewmodel
+
+import androidx.lifecycle.ViewModel
+
+class JobSearchViewModel : ViewModel()
