--- app/src/main/java/ru/practicum/android/diploma/presentation/team/view_model/TeamViewModel.kt
@@ -1,5 +1,0 @@
-package ru.practicum.android.diploma.presentation.team.view_model
-
-import androidx.lifecycle.ViewModel
-
-class TeamViewModel : ViewModel()
