+++ app/src/main/java/ru/practicum/android/diploma/presentation/team/viewmodel/TeamViewModel.kt
@@ -1,0 +1,5 @@
+package ru.practicum.android.diploma.presentation.team.viewmodel
+
+import androidx.lifecycle.ViewModel
+
+class TeamViewModel : ViewModel()
