--- app/src/main/java/ru/practicum/android/diploma/presentation/vacancy/view_model/VacancyViewModel.kt
@@ -1,5 +1,0 @@
-package ru.practicum.android.diploma.presentation.vacancy.view_model
-
-import androidx.lifecycle.ViewModel
-
-class VacancyViewModel : ViewModel()
