+++ app/src/main/java/ru/practicum/android/diploma/presentation/vacancy/viewmodel/VacancyViewModel.kt
@@ -1,0 +1,5 @@
+package ru.practicum.android.diploma.presentation.vacancy.viewmodel
+
+import androidx.lifecycle.ViewModel
+
+class VacancyViewModel : ViewModel()
