--- app/src/main/java/ru/practicum/android/diploma/ui/filtration/choose_country/screen/ChooseCountryScreen.kt
@@ -1,8 +1,0 @@
-package ru.practicum.android.diploma.ui.filtration.choose_country.screen
-
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Modifier
-
-@Composable
-fun ChooseCountryScreen(modifier: Modifier = Modifier) {
-}
