--- app/src/main/java/ru/practicum/android/diploma/ui/filtration/choose_region/screen/ChooseRegionScreen.kt
@@ -1,8 +1,0 @@
-package ru.practicum.android.diploma.ui.filtration.choose_region.screen
-
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Modifier
-
-@Composable
-fun ChooseRegionScreen(modifier: Modifier = Modifier) {
-}
