+++ app/src/main/java/ru/practicum/android/diploma/ui/filtration/country/screen/ChooseCountryScreen.kt
@@ -1,0 +1,8 @@
+package ru.practicum.android.diploma.ui.filtration.country.screen
+
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+
+@Composable
+fun ChooseCountryScreen(modifier: Modifier = Modifier) {
+}
