+++ app/src/main/java/ru/practicum/android/diploma/ui/filtration/industry/screen/IndustrySelectionScreen.kt
@@ -1,0 +1,8 @@
+package ru.practicum.android.diploma.ui.filtration.industry.screen
+
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+
+@Composable
+fun IndustrySelectionScreen(modifier: Modifier = Modifier) {
+}
