--- app/src/main/java/ru/practicum/android/diploma/ui/filtration/industry_selection/screen/IndustrySelectionScreen.kt
@@ -1,8 +1,0 @@
-package ru.practicum.android.diploma.ui.filtration.industry_selection.screen
-
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Modifier
-
-@Composable
-fun IndustrySelectionScreen(modifier: Modifier = Modifier) {
-}
