--- app/src/main/java/ru/practicum/android/diploma/ui/filtration/place_of_work/screen/PlaceOfWorkScreen.kt
@@ -1,8 +1,0 @@
-package ru.practicum.android.diploma.ui.filtration.place_of_work.screen
-
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Modifier
-
-@Composable
-fun PlaceOfWorkScreen(modifier: Modifier = Modifier) {
-}
