+++ app/src/main/java/ru/practicum/android/diploma/ui/filtration/region/screen/ChooseRegionScreen.kt
@@ -1,0 +1,12 @@
+package ru.practicum.android.diploma.ui.filtration.region.screen
+
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.res.stringResource
+import ru.practicum.android.diploma.R
+
+@Composable
+fun ChooseRegionScreen(modifier: Modifier = Modifier) {
+    Text(stringResource(R.string.team))
+}
