+++ app/src/main/java/ru/practicum/android/diploma/ui/filtration/workplace/fragment/PlaceOfWorkFragment.kt
@@ -1,0 +1,28 @@
+package ru.practicum.android.diploma.ui.filtration.workplace.fragment
+
+import android.os.Bundle
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import androidx.compose.ui.platform.ComposeView
+import androidx.compose.ui.platform.ViewCompositionStrategy
+import androidx.fragment.app.Fragment
+import ru.practicum.android.diploma.ui.filtration.workplace.screen.PlaceOfWorkScreen
+import ru.practicum.android.diploma.ui.theme.AppTheme
+
+class PlaceOfWorkFragment : Fragment() {
+    override fun onCreateView(
+        inflater: LayoutInflater,
+        container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View {
+        return ComposeView(requireContext()).apply {
+            setViewCompositionStrategy(ViewCompositionStrategy.DisposeOnViewTreeLifecycleDestroyed)
+            setContent {
+                AppTheme {
+                    PlaceOfWorkScreen()
+                }
+            }
+        }
+    }
+}
