+++ app/src/main/java/ru/practicum/android/diploma/ui/filtration/workplace/screen/PlaceOfWorkScreen.kt
@@ -1,0 +1,8 @@
+package ru.practicum.android.diploma.ui.filtration.workplace.screen
+
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+
+@Composable
+fun PlaceOfWorkScreen(modifier: Modifier = Modifier) {
+}
