--- app/src/main/java/ru/practicum/android/diploma/ui/job_search/fragment/JobSearchFragment.kt
@@ -1,28 +1,0 @@
-package ru.practicum.android.diploma.ui.job_search.fragment
-
-import android.os.Bundle
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
-import androidx.compose.ui.platform.ComposeView
-import androidx.compose.ui.platform.ViewCompositionStrategy
-import androidx.fragment.app.Fragment
-import ru.practicum.android.diploma.ui.job_search.screen.JobSearchScreen
-import ru.practicum.android.diploma.ui.theme.AppTheme
-
-class JobSearchFragment : Fragment() {
-    override fun onCreateView(
-        inflater: LayoutInflater,
-        container: ViewGroup?,
-        savedInstanceState: Bundle?
-    ): View {
-        return ComposeView(requireContext()).apply {
-            setViewCompositionStrategy(ViewCompositionStrategy.DisposeOnViewTreeLifecycleDestroyed)
-            setContent {
-                AppTheme {
-                    JobSearchScreen()
-                }
-            }
-        }
-    }
-}
