--- app/src/main/java/ru/practicum/android/diploma/ui/job_search/screen/JobSearchScreen.kt
@@ -1,8 +1,0 @@
-package ru.practicum.android.diploma.ui.job_search.screen
-
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Modifier
-
-@Composable
-fun JobSearchScreen(modifier: Modifier = Modifier) {
-}
