+++ app/src/main/java/ru/practicum/android/diploma/ui/search/screen/JobSearchScreen.kt
@@ -1,0 +1,8 @@
+package ru.practicum.android.diploma.ui.search.screen
+
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+
+@Composable
+fun JobSearchScreen(modifier: Modifier = Modifier) {
+}
